feat: Refactor filter types and update imports to use custom-advanced-input-filter.type
